requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
sqlalchemy>=1.4
pydantic>=1.10
python-jose>=3.3
passlib[bcrypt]>=1.7
psycopg2-binary>=2.9 ; platform_system == "Windows"
